+++ ums-api/src/main/java/org/ums/fee/FeeManager.java
@@ -2,5 +2,5 @@ package org.ums.fee;
 
 import org.ums.manager.ContentManager;
 
-public interface FeeManager extends ContentManager<Fee, MutableFee, Integer> {
+public interface FeeManager extends ContentManager<Fee, MutableFee, Long> {
 }
