+++ ums-core/src/main/java/org/ums/fee/FeeDaoDecorator.java
@@ -2,6 +2,6 @@ package org.ums.fee;
 
 import org.ums.decorator.ContentDaoDecorator;
 
-public class FeeDaoDecorator extends ContentDaoDecorator<Fee, MutableFee, Integer, FeeManager>
+public class FeeDaoDecorator extends ContentDaoDecorator<Fee, MutableFee, Long, FeeManager>
     implements FeeManager {
 }
